requirements.txt
streamlit>=1.35.0
openai>=1.40.0
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
transformers>=4.45.0,<5
pypdfium2>=4.30.0
tiktoken>=0.7.0
numpy>=1.26.4
imageio-ffmpeg>=0.6.0
pydantic>=2.8.2
pytest>=8.2.0
pytest-mock>=3.14.0
pandas>=2.2.2
